requirements.txt
asgiref==3.5.2
attrs==22.1.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2022.6.15
chardet==4.0.0
click==8.0.3
cloud-init==22.3.4
colorama==0.4.5
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu1
Django==4.1.4
django-anymail==8.6
django-countries==7.4.2
django-taggit==3.1.0
filelock==3.8.2
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.8.0
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.6.0
keyring==23.8.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pexpect==4.8.0
Pillow==9.3.0
pip==22.2
platformdirs==2.6.0
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyGObject==3.42.2
PyHamcrest==2.0.2
PyJWT==2.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.3.0+ubuntu2
python-debian==0.1.46
python-magic==0.4.26
pytz==2022.2.1
PyYAML==5.4.1
requests==2.27.1
SecretStorage==3.3.3
service-identity==18.1.0
setuptools==59.6.0
six==1.16.0
sos==4.4
sqlparse==0.4.3
ssh-import-id==5.11
systemd-python==235
Twisted==22.4.0
typing_extensions==4.4.0
ubuntu-advantage-tools==27.11.2
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.9
virtualenv==20.17.1
wadllib==1.3.6
wheel==0.37.1
zipp==1.0.0
zope.interface==5.4.0
